requirements.txt
torch 
torchvision 
torchaudio
pandas
scikit-learn
scipy
numpy protobuf==3.16.0
onnx
onnxruntime
tensorflow-addons
Pillow